+++ client/src/hooks/mutations/accounts/useRestoreAccountMutation.ts
@@ -1,0 +1,35 @@
+import { notifications } from "@mantine/notifications";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { AxiosError } from "axios";
+import {
+  accountsQueryKey,
+  institutionsQueryKey,
+  transactionsQueryKey,
+  translateAxiosError,
+} from "~/helpers/requests";
+import { useAuth } from "~/providers/AuthProvider/AuthProvider";
+
+export const useRestoreAccountMutation = () => {
+  const queryClient = useQueryClient();
+  const { request } = useAuth();
+
+  return useMutation({
+    mutationFn: async (accountId: string) =>
+      await request({
+        url: `/api/account/restore`,
+        method: "POST",
+        params: { guid: accountId },
+      }),
+    onSuccess: async () => {
+      await queryClient.invalidateQueries({ queryKey: [accountsQueryKey] });
+      await queryClient.invalidateQueries({ queryKey: [institutionsQueryKey] });
+      await queryClient.invalidateQueries({ queryKey: [transactionsQueryKey] });
+    },
+    onError: (error: AxiosError) => {
+      notifications.show({
+        color: "var(--button-color-destructive)",
+        message: translateAxiosError(error),
+      });
+    },
+  });
+};
